hack(scim): Negotiate OAuth2 access tokens out of oauth2 config in SCIM
This hack expects on the SCIM provider in the token field a needle
(`__OAUTH2__`), when detected it'll read the rest of the value as a
URL-encoded query string.

Like:

`__OAUTH2__{url-encoded-values}`.

And `url-encoded-values` must have:

- `auth_url`: The URL where to get the `access_token`
- `client_id`: The OAuth2 `client_id`
- `client_secret`:  The OAuth2 `client_secret`
- `grant_type`: The OAuth2 `grant_type The OAuth2 `client_id`.`. This
  hack expects ANYTHIGN that makes the OAuth2 provider return an
  `access_token` in the top-level key of the response. Salesforce does
  it with `client_credentials`
- `expiry` a timedelta expressing how long should the negotiated token
  be kept in cache, it defaults to 1h.

With these parameters, the SCIM client negotiates an OAuth2 bearer
token and caches it (for the time dictated by `expiry`), and uses it
in follow up calls.

[0]: https://help.salesforce.com/s/articleView?id=xcloud.remoteaccess_oauth_client_credentials_flow.htm&type=5

Signed-off-by: Jose D. Gomez R. <jose.gomez@suse.com>